KD8ZRC.Mapview.Execute

Description

Provides a function that kicks off the backend (TODO: better description once we have a better idea of how this will actually work).

Documentation

mapview :: MapviewConfig t -> IO ()

Takes a MapviewConfig and uses its _mvDownlinkSpawn callback to start our backend process.

TODO: Threading stuff -- where, when, how?