KD8ZRC.Mapview.Types

Description

Provides fundamental types intrinsic to the rest of the mapview library. This shouldn't contain types specific to any particular use of the library (suggested types can be found in the KD8ZRC.Mapview.Utility.* set of modules).

Documentation

type MV t a = ReaderT (MapviewConfig t) IO a

The MV type is a monad transformer which carries around our MapviewConfig, making it accessible to all callback hooks.

newtype PacketLineCallback t

The type for callbacks which occur immediately after a packet is received, before even parsing occurs.

Constructors

PacketLineCallback

Fields

getPacketLineCallback :: ByteString -> MV t ()

data ParsedPacketCallback t

The type for callbacks which occur immediately after a packet is parsed.

Constructors

ParseSuccessCallback

Fields

getParseSuccessCallback :: t -> MV t ()

ParseFailureCallback

Fields

getParseFailureCallback :: Doc -> MV t ()

data MapviewConfig t

The configuration for this instance of Mapview. The t parameter is the type that the telemetry parser parses into, if it is successfully able to parse the data.

Constructors

MapviewConfig

Fields

_mvParser :: forall m. (Monad m, DeltaParsing m, Errable m) => m t

Determines how a packet is parsed.

_mvPacketLineCallback :: [PacketLineCallback t]

Determines what to do immediately after a packet is received, before it is parsed.

_mvDownlinkSpawn :: MV t ()

As soon as Mapview is launched, this callback is run and should begin the process of listening for telemetry data.

_mvParsedPacketCallback :: [ParsedPacketCallback t]

These callbacks get called depending on whether or not a parse was successful.

mvParser :: forall t m. (Monad m, DeltaParsing m, Errable m) => Getter (MapviewConfig t) (m t)

mvParsedPacketCallback :: forall t. Lens' (MapviewConfig t) [ParsedPacketCallback t]

mvPacketLineCallback :: forall t. Lens' (MapviewConfig t) [PacketLineCallback t]

mvDownlinkSpawn :: forall t. Lens' (MapviewConfig t) (MV t ())