KD8ZRC.Mapview.Utility.CRC

Description

Provides standard implementations of CRC utility functions.

Documentation

newtype TelemetryCRC a

The CRC code that we received in the telemetry message from the downlink.

Constructors

TelemetryCRC a

newtype CalculatedCRC a

The CRC code that we calculated for the message we received in the downlink.

Constructors

CalculatedCRC a

data CRCConfirmation a

The result after checking a message for a valid CRC.

Constructors

CRCMatch

Fields

_crcMatchValue :: a

CRCMismatch

Fields

_crcTelemetryValue :: TelemetryCRC a

_crcCalculatedValue :: CalculatedCRC a

validateCRC :: Eq a => TelemetryCRC a -> CalculatedCRC a -> CRCConfirmation a

crcIsMatch :: CRCConfirmation a -> Bool