KD8ZRC.Mapview.Utility.Downlink

Description

Provides standard implementations of raw downlink callbacks, including:

How to receive the downlink (modem commands, etc.)

Standard callbacks for parsing, logging, and other data manipulation happens in other modules. This module only deals with how to obtain the downlink data, by e.g., calling out to minimodem or fldigi-shell, etc.

Documentation

packetCallbackCaller :: ByteString -> MV t ()

Fires off the callbacks in our MapviewConfig's mvPacketLineCallback field. For now, you probably want to make sure to call this from your downlink-obtaining function, on each line.

parsedPacketCallbackCaller :: ByteString -> MV t ()

Fires off the callbacks in our MapviewConfig's mvParsedPacketCallback field. For now, you probably want to make sure to call this from your downlink-obtaining function, on each line. This should be fixed at some point, because it forces us to give up a separation of concerns, which is annoying. I'm not yet sure of the correct abstraction, however.

modemStdout

Arguments

:: Text
The modem command to run (e.g. minimodem or fldigi-shell)

-> [Text]
Flags/arguments to pass to the modem command

-> MV t ()

This callback provides a way to obtain downlink data by shelling out to an audio modem implementation, such as minimodem or fldigi, and using each (newline-separated) line of its standard output as a downlink packet.